the great deshittification
i've learned a lot about embedding languages in odin recently. this is the epic cleanup this project needed

- completely remove macros.odin, no more manual implementation (poorly) of macros
- instead we have macros.c, a thin wrapper that just exposes the necessary macros
- this means we can drop tons of struct definitions (nearly all)!
- it means we have way fewer weird bugs and edge cases
- it means performance is way up

bonus:

- implement FSM and Animation natively
- needs some cleanup still

+++ source/callbacks.odin
@@ -50,7 +50,7 @@ determine_game_callbacks :: proc() {
 			log.warnf("Unable to confirm existence of `%s` callback", callback.method_name)
 		}
 
-		arity := mrb.integer_p(arity_result) ? mrb.integer(arity_result) : -1
+		arity := mrb.integer_p(arity_result) ? i32(mrb.integer(arity_result)) : -1
 
 		if arity == -1 {
 			callback.has_flag^ = false
